refactor: move factories inside filters
